ecomm_bot/requirements.txt
flask 
flask-cors 
google-generativeai
